2D FDTD calculated and implemented
I calculated the equations for the FDTD in 2D from the Maxwell equations, which was surprisingly simple and figured the vectorization out to compute it efficently. The only thing that is missing is the boundary conditions. I dont want to use Mur Boundary conditions, because I already thought about it maybe having problems with corners and reflections at a circle.